Improvements
- Fixed some log files not being cleaned up
- Added back next and infinity view to pagination code
- Additional Found server services are now found on the homepage
- Additional server services are now checked for with server paths
- Added lots of server services to check for
- Created server services widget
- Raised index limit to 800 items in new indexes
- kodi2web now supports updating audio and video libraries on remote kodi clients
- parallel options now only use half the available CPUs in order to improve availability of other server services
- kodi2web now runs for video items at the end of processing nfo2web
- kodi2web now runs for audio items at the end of processing music2web
- hls.js now only updates on build if another version has not been downloaded into the build directory
- merged random and new playlists into the playlists header button
- Added icon to weather homepage entry
- Added icon to homepage fortune
- Redesigned how wiki thumbnails and index entries
- Added styles to make wiki formatting look better

+++ makefile
@@ -73,7 +73,7 @@ build: build-deb
 	# install the build tools
 	sudo make build-deb;
 upgrade-hls: node_modules/hls.js/dist/hls.js
-	npm install --save hls.js
+	if ! test -f node_modules/hls.js/dist/hls.js;then npm install --save hls.js;fi
 screenshots:
 	# create a series of screenshots from the currently installed version
 	mkdir -p screenshots/
